Refactor structure of service folder
