+++ app/service/user_service.py
@@ -53,20 +53,13 @@ def create_user(username, password, user_type, email=None, address=None):
         raise e
 
 def verify_password(user, password):
-    """驗證用戶密碼"""
+    """
+    驗證用戶密碼
+    
+    :param user: dict, user data
+    :param password: str, password to verify
+    :return: bool, whether the password is correct
+    """
     if user and check_password_hash(user['password'], password):
         return True
-    return False 
-
-def get_customer_profile_by_username(username):
-    """根據用戶名獲取完整的顧客資料（包含用戶基本資料和顧客特定資料）"""
-    db = get_db()
-    # 連接 users 表和 customer 表獲取完整資訊
-    profile = db.execute(
-        'SELECT u.user_id, u.name, u.created_at, c.email, c.address, c.balance '
-        'FROM users u '
-        'LEFT JOIN customer c ON u.user_id = c.user_id '
-        'WHERE u.name = ? AND u.user_type = "customer"',
-        (username,)
-    ).fetchone()
-    return profile
+    return False
